Added coverage to tests
