+++ package.json
@@ -4,7 +4,7 @@
   "description": "Small library for .",
   "main": "./lib/index.js",
   "scripts": {
-    "test": "mocha --recursive",
+    "test": "nyc --reporter=html --reporter=text mocha",
     "test:watch": "npm test -- --watch",
     "cover": "cross-env BABEL_ENV=commonjs istanbul cover _mocha -- --compilers js:babel-register --recursive"
   },
@@ -32,7 +32,8 @@
   "devDependencies": {
     "chai": "^4.1.2",
     "istanbul": "^1.0.0-alpha",
-    "mocha": "^4.0.1"
+    "mocha": "^4.0.1",
+    "nyc": "^11.2.1"
   },
   "dependencies": {
     "lodash.get": "^4.4.2"
